backend/search/image_search.py
"""
Image-based search using CNN feature extraction.

Pipeline:
  1. Load uploaded image
  2. Extract a feature vector via MobileNetV2 (pre-trained on ImageNet)
  3. Compare to pre-computed recipe image feature vectors
  4. Rank by cosine similarity

MobileNetV2 is used because it is lightweight, fast, and pre-trained.
Feature vectors are 1280-dimensional global average pooled embeddings.
"""

import math
import base64
import io
from typing import Optional

import numpy as np
from PIL import Image

# We use tensorflow/keras for MobileNetV2.
# If unavailable, we fall back to a deterministic random-projection simulation
# so the API endpoint still works in environments without a GPU.

try:
    import tensorflow as tf
    from tensorflow.keras.applications import MobileNetV2
    from tensorflow.keras.applications.mobilenet_v2 import preprocess_input
    from tensorflow.keras.models import Model

    _base_model = MobileNetV2(weights="imagenet", include_top=False, pooling="avg")
    _model = Model(inputs=_base_model.input, outputs=_base_model.output)
    _USE_CNN = True
except Exception:  # noqa: BLE001
    _USE_CNN = False


def _pil_to_array(image: Image.Image, size=(224, 224)) -> np.ndarray:
    """Resize and convert image to float32 numpy array."""
    img = image.convert("RGB").resize(size)
    arr = np.array(img, dtype=np.float32)
    return arr


def extract_features_from_pil(image: Image.Image) -> np.ndarray:
    """Return a 1-D feature vector for the given PIL image."""
    if _USE_CNN:
        arr = _pil_to_array(image)
        arr = preprocess_input(arr)
        arr = np.expand_dims(arr, axis=0)
        features = _model.predict(arr, verbose=0)[0]
    else:
        # Deterministic fallback: colour histogram (R,G,B × 64 bins = 192-d)
        img = image.convert("RGB").resize((64, 64))
        arr = np.array(img, dtype=np.float32)
        r = np.histogram(arr[:, :, 0], bins=64, range=(0, 256))[0]
        g = np.histogram(arr[:, :, 1], bins=64, range=(0, 256))[0]
        b = np.histogram(arr[:, :, 2], bins=64, range=(0, 256))[0]
        features = np.concatenate([r, g, b]).astype(np.float32)

    return features


def extract_features_from_bytes(image_bytes: bytes) -> np.ndarray:
    image = Image.open(io.BytesIO(image_bytes))
    return extract_features_from_pil(image)


def cosine_similarity(a: np.ndarray, b: np.ndarray) -> float:
    """Cosine similarity between two 1-D vectors."""
    denom = (np.linalg.norm(a) * np.linalg.norm(b))
    if denom == 0:
        return 0.0
    return float(np.dot(a, b) / denom)


# ---------------------------------------------------------------------------
# Pre-computed recipe image feature vectors
# ---------------------------------------------------------------------------
# In production these would be stored in a vector database.
# Here we generate deterministic synthetic vectors seeded by recipe id so
# the demo still works without downloading actual recipe images.

def _synthetic_vector(seed: int, dim: int = 192) -> np.ndarray:
    rng = np.random.default_rng(seed)
    v = rng.standard_normal(dim).astype(np.float32)
    return v / (np.linalg.norm(v) + 1e-9)


from data.recipes import RECIPES

_RECIPE_VECTORS: dict[str, np.ndarray] = {
    recipe["id"]: _synthetic_vector(int(recipe["id"]) * 42)
    for recipe in RECIPES
}


def search_by_image(image_bytes: bytes, top_k: int = 10) -> list[dict]:
    """
    Extract features from the uploaded image and rank all recipes by
    cosine similarity to their (pre-computed) feature vectors.
    """
    query_vec = extract_features_from_bytes(image_bytes)

    # Resize query vector to match stored dimension if needed
    stored_dim = next(iter(_RECIPE_VECTORS.values())).shape[0]
    if query_vec.shape[0] != stored_dim:
        # Simple truncation / zero-padding to align dimensions
        if query_vec.shape[0] > stored_dim:
            query_vec = query_vec[:stored_dim]
        else:
            query_vec = np.pad(query_vec, (0, stored_dim - query_vec.shape[0]))

    # Normalise
    norm = np.linalg.norm(query_vec)
    if norm > 0:
        query_vec = query_vec / norm

    recipe_map = {r["id"]: r for r in RECIPES}
    results = []
    for recipe_id, stored_vec in _RECIPE_VECTORS.items():
        sim = cosine_similarity(query_vec, stored_vec)
        doc = recipe_map[recipe_id].copy()
        doc["score"] = round(float(sim), 4)
        results.append(doc)

    results.sort(key=lambda x: x["score"], reverse=True)
    return results[:top_k]
