backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
numpy>=1.26.0
Pillow>=10.3.0
rapidfuzz>=3.9.0
# Optional – remove if you do not need real CNN features:
# tensorflow>=2.16.0
